+++ example/lib/config.dart
@@ -1,2 +1,0 @@
-const String apiKey = 'abc';
-const String measurementId = '';