+++ lib/src/ambilytics.dart
@@ -53,14 +53,21 @@ void resetInitialized() {
 // Challenge is tha it can be done for all custom events, not sure about standard (e.g. screen_view)
 
 /// Prepares analytics for usage. Doesn't throw errors, in debug mode throws assertions. If Ambilytics fails to initialize [isAmbilyticsInitialized] returns false.
+///
 /// If the platform is Android, iOS, macOS, or Web, Firebase Analytics will be used ([_firebaseAnalytics] instance will be initialized).
 /// Otherwise, GA4 Measurement protocol and custom events will be used ([_ambilytics] instance will be initialized).
+///
 /// If [fallbackToMP] is true, than Measurement Protocol will be used if Firebase analytics fails to initialize. E.g. you can skip configuring Firebase Analytics in native projects and use MP for all platforms.
+///
 /// If [disableAnalytics] is `true`, analytics will not be initialized, any analytics calls will be ingonred,
 /// [_firebaseAnalytics] and [_ambilytics] instances will be null. Usefull for the scanarious when toy wish to disable analytics.
+///
 /// If [sendAppLaunch] is true, "app_launch" will be sent with "platfrom" param value corresponding runtime platform (i.e. Windows)
+///
 /// [firebaseOptions] forwards options (e.g. generated via `flutterfire configure`) to `Firebase.initializeApp()`.
+///
 /// [apiSecret] and [measurementId] must be set in order to enable GA4 Measurement protocol and have [_ambilytics] initialized.
+///
 /// [userId] allows overriding user identifier. If not provided, default user ID will be used by Firebase Analytics OR
 /// or a GUID will be created and put to shared_preferences storage (for Windows and Linux).
 Future<void> initAnalytics(
@@ -116,8 +123,8 @@ Future<void> initAnalytics(
       }
     }
     if (measurementId != null && apiSecret != null) {
-      _ambilytics = AmbilyticsSession(measurementId, apiSecret,
-          'test_user_${defaultTargetPlatform.name}', false);
+      _ambilytics =
+          AmbilyticsSession(measurementId, apiSecret, ambiUserId, false);
     }
     if (_ambilytics != null || _firebaseAnalytics != null) {
       _initialized = true;